requirements.txt
httpx>=0.27.0
msal>=1.29.0
openai>=1.40.0
pydantic>=2.8.0
python-dotenv>=1.0.0
tenacity>=9.0.0

# Dashboard
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
anthropic>=0.40.0
python-multipart>=0.0.9
